+++ src/components/popup-header/index.js
@@ -2,22 +2,24 @@ import React from 'react'
 import classNames from 'classnames'
 import './index.less'
 
+const noop = ()=>{}
+
 var PopupHeader = React.createClass({
   render(){
-    const {leftText, rightText, title, showBottomBorder, className, ...others} = this.props;
+    const {leftText, rightText, title, showBottomBorder, onClickRight, onClickLeft, className, ...others} = this.props;
 
     const cls = classNames('mt-popup-header', className);
 
     return (
       <div className={ cls }>
-        <div className="mt-popup-header-left" onClick={ this.onClickRight }>
-          {{ leftText }}
+        <div className="mt-popup-header-left" onClick={ onClickLeft }>
+          { leftText }
         </div>
         <div className="mt-popup-header-title">
-          {{ title }}
+          { title }
         </div>
-        <div className="mt-popup-header-right" onClick={ this.onClickLeft }>
-          {{ rightText }}
+        <div className="mt-popup-header-right" onClick={ onClickRight }>
+          { rightText }
         </div>
       </div>
     );
@@ -30,6 +32,8 @@ PopupHeader.propTypes = {
   rightText: React.PropTypes.string,
   title: React.PropTypes.string,
   showBottomBorder: React.PropTypes.bool,
+  onClickRight: React.PropTypes.func,
+  onClickLeft: React.PropTypes.func,
 };
 
 PopupHeader.defaultProps = {
@@ -37,6 +41,8 @@ PopupHeader.defaultProps = {
   rightText: '',
   title: '',
   showBottomBorder: true,
+  onClickRight: noop,
+  onClickLeft: noop,
 };
 
 export default PopupHeader